refactor: 拆分agent快照窗口runner逻辑
